Update CMD to remove sudo and use rbash
The default command should be usable if someone
just does a naive docker run command. Now that
we have removed the ONBUILD command and the 
USER airstack from the end of the Dockerfile,
we can have a simplified CMD that starts up the
programs and then runs bash in the foreground. 
Rbash is just a restricted form of bash that 
seemed like a decent default shell choice.